# docs/openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AccountPatchRequest" : {
        "properties" : {
          "locale" : {
            "enum" : [ "ca", "es", "en" ],
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiError" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "additionalProperties" : true,
            "type" : "object"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "details", "message", "traceId" ],
        "type" : "object"
      },
...
      "Me" : {
        "description" : "R-01-15 app bootstrap. Membership is absent without club context; features are branding.modules.",
        "properties" : {
          "account" : {
            "$ref" : "#/components/schemas/AccountSummary"
          },
          "features" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "impersonation" : {
            "$ref" : "#/components/schemas/Impersonation"
          },
          "membership" : {
            "$ref" : "#/components/schemas/MembershipSummary"
          }
        },
        "required" : [ "account", "features" ],
        "type" : "object"
      },
...
      "OnboardingRequest" : {
        "properties" : {
          "imageConsent" : {
            "type" : "boolean"
          },
          "locale" : {
            "enum" : [ "ca", "es", "en" ],
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "phone" : {
            "description" : "Requested only when configured by the club",
            "type" : "string"
          },
          "privacyAccepted" : {
            "type" : "boolean"
          },
          "privacyPolicyVersion" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "privacyAccepted", "privacyPolicyVersion" ],
        "type" : "object"
      },
      "OnboardingState" : {
        "description" : "Pending onboarding, requested data fields and the current privacy policy; S01 §14",
        "properties" : {
          "locale" : {
            "enum" : [ "ca", "es", "en" ],
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "onboardingPending" : {
            "type" : "boolean"
          },
          "phone" : {
            "type" : "string"
          },
          "privacyPolicyUrl" : {
            "format" : "uri",
            "type" : "string"
          },
          "privacyPolicyVersion" : {
            "type" : "string"
          },
          "requestedFields" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "onboardingPending", "privacyPolicyUrl", "privacyPolicyVersion", "requestedFields" ],
        "type" : "object"
      },
...
      "ProfileRequest" : {
        "properties" : {
          "activeProfile" : {
            "$ref" : "#/components/schemas/Profile"
          },
          "remember" : {
            "type" : "boolean"
          }
        },
        "required" : [ "activeProfile", "remember" ],
        "type" : "object"
      },
      "ProfileResponse" : {
        "properties" : {
          "access_token" : {
            "type" : "string"
          }
        },
        "required" : [ "access_token" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "AgilityHub Core API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/me" : {
      "get" : {
        "description" : "Any valid account token. R-01-15: only the current host/JWT membership, role profiles and enabled features. The E0 club bootstrap works; global account and impersonation bootstrap are completed in E1.",
        "operationId" : "me",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Me"
                }
              }
            },
            "description" : "App bootstrap (Me)"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Get the current account and active club context",
        "tags" : [ "identity" ]
      },
      "patch" : {
        "description" : "Any valid account token. R-01-14.",
        "operationId" : "update",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AccountPatchRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Me"
                }
              }
            },
            "description" : "Updated app bootstrap (Me)"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "LOCALE_NOT_SUPPORTED"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Update account language or display name",
        "tags" : [ "identity" ]
      }
    },
    "/api/v1/me/onboarding" : {
      "get" : {
        "description" : "Any valid account token. S01 §14, T-01-26. Implemented in E1-T06.",
        "operationId" : "onboarding",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingState"
                }
              }
            },
            "description" : "Pending onboarding and current privacy policy"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Get required first-access profile and privacy steps",
        "tags" : [ "identity" ]
      },
      "put" : {
        "description" : "Any valid account token. S01 §14, T-01-26: privacyAccepted must be true for the current policy version. Profile data may be deferred; imageConsent is optional. Implemented in E1-T06.",
        "operationId" : "completeOnboarding",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OnboardingRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingState"
                }
              }
            },
            "description" : "Updated onboarding state"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Complete first-access privacy consent and optional profile fields",
        "tags" : [ "identity" ]
      }
    },
    "/api/v1/me/password" : {
      "put" : {
        "description" : "Account token, never impersonated. R-01-05: current is required only when a password already exists.",
        "operationId" : "password",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Password changed"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "PASSWORD_TOO_SHORT, PASSWORD_MISMATCH, PASSWORD_COMPROMISED"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "INVALID_CREDENTIALS for current password"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "FORBIDDEN for impersonated accounts"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Set or change the current account password",
        "tags" : [ "identity" ]
      }
    },
    "/api/v1/me/profile" : {
      "put" : {
        "description" : "Account token with club context. R-01-07. Returns a fresh access token.",
        "operationId" : "profile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description" : "New access token"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "PROFILE_NOT_AVAILABLE"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Select an available club profile",
        "tags" : [ "identity" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "url" : "https://core.agilitydoghub.com"
  }, {
    "url" : "http://localhost:8080"
  } ]
}